move more stuff out of doc elements into annotations; fix some typing bugs
